requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
httpx==0.25.1
python-dotenv==1.0.0
bleach==6.1.0
slowapi==0.1.8
pytest==7.4.3
requests==2.31.0
Jinja2==3.1.2
google-generativeai==0.8.3